+++ src/strata/validators/policies/policy_engine.py
@@ -62,6 +62,7 @@ class PolicyEngine:
 
     def _create(self, policy_model: PolicyModel) -> BasePolicy:
         """Dispatch policy type to its concrete implementation."""
+        from strata.validators.policies.checkov_policy import CheckovPolicy
         from strata.validators.policies.cost_threshold_policy import CostThresholdPolicy
         from strata.validators.policies.cve_max_severity_policy import CveMaxSeverityPolicy
         from strata.validators.policies.naming_policy import NamingPolicy
@@ -91,6 +92,7 @@ class PolicyEngine:
             "sbom_license": SbomLicensePolicy,
             "cve_max_severity": CveMaxSeverityPolicy,
             "cost_threshold": CostThresholdPolicy,
+            "checkov": CheckovPolicy,
             "path_convention": PathConventionPolicy,
         }